+++ CHANGELOG.md
@@ -7,11 +7,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### ✅ Added
 - Screensharing from iOS devices
 - Remote pinning of users
-- Add XCPrivacy manifest [#139](https://github.com/GetStream/stream-chat-swift/pull/139)
+- Add XCPrivacy manifest
+- Custom Audio Filters
 
 ### 🔄 Changed
 - Factory method for creating `VideoCallParticipantView`
 - `VideoCallParticipantView` init method
+
+### 🐞 Fixed
+- Stability improvements
+- CPU usage improvements
 
 # [0.2.0](https://github.com/GetStream/stream-video-swift/releases/tag/0.2.0)
 _July 18, 2023_